requirements.txt
django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-allauth==0.61.1
psycopg2-binary==2.9.9
google-generativeai==0.5.4
yfinance==0.2.37
python-decouple==3.8
Pillow==10.3.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
